Switch from ed25519 to ecdsa (#52)
* Switch from ed25519 to ecdsa

ed25519 hasn't yet been approved and won't work with most KMS systems, so we need to use ecdsa until it will.

I chose the P521 elliptic curve because it is the most secure one available, although I believe the tradeoff is larger keys.
It might be worth it to choose a different curve if we want people to be able to pass in public keys via command line.

Signed-off-by: Priya Wadhwa <priyawadhwa@google.com>

* Use pointer for pubilc key since rekor expects that

Signed-off-by: Priya Wadhwa <priyawadhwa@google.com>

* Hash payload before using SignASN1 or VerifyASN1

Signed-off-by: Priya Wadhwa <priyawadhwa@google.com>

* use pointer

Signed-off-by: Priya Wadhwa <priyawadhwa@google.com>

* Switch to P256 curve so that this will work with KMS's

Signed-off-by: Priya Wadhwa <priyawadhwa@google.com>
Signed-off-by: Dan Lorenc <dlorenc@google.com>