Allow passing in additional annotations to check for in deduper
The TestTlog test was failing because it was going through this flow:
1. Sign and verify image
2. Turn on experimental feature
3. Sign and verify image again

The second verification was failing with bundling because the deduper thought the image already existed, even though we were trying to tack on a new bundle annotation.

This way, we can pass in extra annotations we want the deduper to check for before deciding the signature alrady exists.

Signed-off-by: Priya Wadhwa <priyawadhwa@google.com>