add flags to specify slot, PIN, and touch policies for security keys
This permits flexibility to use different PIN and touch policies
for container signing. Key generation will use the default PIN
and touch policies for the specified slot unless additional flags
are set to attempt to override the default. The security key is
not guaranteed to honor a particular PIN and touch policy for the
specified slot.

Signed-off-by: Dino A. Dai Zovi <ddz@theta44.org>