pivkey: Refactor
This refactor abstracts away the need to directly call the upstream
piv-go library, implements a common SignerVerifier interface so there
is no need to maintain separate types for ECDSA and RSA private keys,
and overall is more idiomatic. Additionally, it fixes an issue where the
previous initialization functions would lock access to the PIV token,
preventing subsequent access even if the originall caller no longer had
a need to access the token.

Signed-off-by: James Alseth <james@jalseth.me>