Break off a `fulcioroot` package.
The `cosigned` webhook pulls in the Fulcio roots, and runs as a K8s controller, which consumes `klog`.  However, some of the certificate transparency stuff the Fulcio package pulls in consumes `glog`.  These both define conflicting `-log_dir` flags, which cause `cosigned` to crash on startup.

With this change, `cosigned` can use `fulcioroots.Get` to load the roots without pulling in `glog`.

In a subsequent change, I have tests that should catch this before a breaking change merges.

Signed-off-by: Matt Moore <mattomata@gmail.com>