Split apart `fulcioverifier` for transparency log verification.
Split off a new constructor for the fulcio signer that performs transparency log verification.  This simplifies the tests a bit because they no longer need to mock out `VerifySCT`.  However, the main purpose of this change is to enable folks to vendor `fulcio.NewSigner` without pulling in ~600k LoC of dependencies, since the CT stuff pulls in trillian, prometheus, envoy's go controlplane stuff, ...

Signed-off-by: Matt Moore <mattomata@gmail.com>