Add a simple "DO NOT SUBMIT" workflow.
This is a workflow I wrote for Knative, which is based on a presubmit check we had at Google.

Essentially, it fails the check (and flags the line) if it sees the words `DO NOT SUBMIT`.

So a comment practice with "work in progress" changes is to tag things that shouldn't get merged with:
```go
// TODO(mattmoor): DO NOT SUBMIT
```

Personally, I find this kind of workflow keeps me honest about things, so I thought I'd share!

Signed-off-by: Matt Moore <mattomata@gmail.com>