Start to flesh out `Signatures` and `internal/oci/remote`
Starting with `cosign.FetchSignaturesForImageDigest`, I am starting to peel out some of the existing logic, and try to define useful interfaces within `internal/oci.Signatures` and implement them in `internal/oci/remote.Signatures`.

This is still somewhat rough in parts, so I suspect that as we build it up, we will iterates on parts of it.

Signed-off-by: Matt Moore <mattomata@gmail.com>