Start to sketch `internal/oci/remote.SignedImage[Index]`.
This follows the pattern GGCR uses for functional `remote.Options`, and starts to define `SignedImage[Index]` methods, which build upon the `Signatures` implementation from a prior change.

This also tries to start adopting this pattern a few places, which drops bits of boilerplate (or changes them to functional options).

As part of this, I'm combining `FetchSignaturesForImage[Digest]` and tweaking it's signature a bit (no pun intended!), but I think long term we'll just want folks to consume the `remote` package directly themselves.

Signed-off-by: Matt Moore <mattomata@gmail.com>