Rename `FetchSignaturesForImage` to `Reference` (#715)
This is sort of a throwaway, since I want to kill this method, but I realized staring at it that it's a misnomer since it actually needs to handle `ImageIndex` as well today.  The name misled me into using `ociremote.SignedImage` here where I really wanted `ociremote.SignedEntity`.  Anyways, VS Code made it easy to change, so I changed it :D

Signed-off-by: Matt Moore <mattomata@gmail.com>