Start to build up a `mutate` package. (#729)
* Start to build up a `mutate` package.

This is not yet used, but start to lay out some useful utilities for walking over and manipulating Images and Indices.

The first bit is `mutate.AppendManifests`, which builds around the utility with the same name in GGCR.  This utility produces an `oci.SignedImageIndex`, where (because of the mutation) the signatures are necessarily empty, but which provides access to the signatures of the entities contained within it.  This was mostly useful for constructing images for tests.

The second bit is `mutate.Map`, which surfaces a way to apply a `Mutator` function to the `oci.SignedEntity`s contained within a particular `oci.SignedEntity`.  Notable features:
 * `Mutator`s can control whether they descend into the children of indices (useful for recursive signing or signature verification).
 * `Mutator`s can filter entities by returning `nil` without `error`.
 * `Mutator`s are called before and (if changed) after its children are walked, which is detectable via helpers on `ctx`.
 * `Mutator`s can return their input entity if they are just a simple readonly walk.

Signed-off-by: Matt Moore <mattomata@gmail.com>

* Inline GGCR interface, presere annotations.

Signed-off-by: Matt Moore <mattomata@gmail.com>