Switch most things to use `oci.Signature` directly.
This changes most of the callers of `FetchSignaturesForReference` to use `ociremote.SignedEntity` directly, and eliminates essentially all usage of `SignedPayload` except one specific use case.

The places that remain calling `FetchSignaturesForReference` and using `SignedPayload` are marshalling the `SignedPayload` (one is in the cosign CLI, the other is copasetic and has rego references).

Related: https://github.com/sigstore/cosign/issues/666
Signed-off-by: Matt Moore <mattomata@gmail.com>