Eliminate `AttachedImageTag` in favor of `ociremote` variants.
We were starting to accumulate a large number of variants of this method, which all do roughly equivalent things.

This change consolidates around a handful of methods in the `ociremote` package oriented around the kind of attachment they represent.
 * `ociremote.SignatureTag(ref, opts...)` produces the tag for referencing attached signatures.
 * `ociremote.AttestationTag(ref, opts...)` produces the tag for referencing attached attestations.
 * `ociremote.SBOMTag(ref, opts...)` produces the tag for referencing attached SBOMs.

These share some core logic to be efficient, but they:
1. Check whether they were passed a digest,
2. If not, resolves the digest using the supplied options,
3. Construct the final tag based on the options (e.g. target repo) and the resolved or passed digest.

Related: https://github.com/sigstore/cosign/issues/666
Signed-off-by: Matt Moore <mattomata@gmail.com>