Add `Annotations` to `oci.Signature`.
This change has two main parts, the first is largely as the title indicates, which is to add an `Annotations()` method to `oci.Signature` (and our implementations thereof).  This is largely motivated by the second half of this change.

The second part of this change is starting to orient `UploadSignature` to operate in terms of our `oci` interfaces.  For a much longer-form version of the direction I'd like to drive this towards, see here: https://github.com/sigstore/cosign/issues/666#issuecomment-923609067.  This is a bit messier than I'd like, but I think it is progress towards the final form outlined in the issue since it starts to orient things like deduplication around the `oci.Signature` interface.

Related: https://github.com/sigstore/cosign/issues/666
Signed-off-by: Matt Moore <mattomata@gmail.com>