Move the `empty.Signatures()` on 404 semantic into the lib.
The more I think about this semantics, the more I think it makes sense for it to be a native part of the library.

In places that are dealing with `SignedEntity` today, we more or less expect this semantic when we access their `Signatures`, and other places sort of do this already (e.g. `mutate.AppendManifests` implements `Signatures()` by returning `empty.Signatures()`).

Hopefully this will improve the DevX of the library a bit.

Related: https://github.com/sigstore/cosign/issues/666
Signed-off-by: Matt Moore <mattomata@gmail.com>