Share more of the tlog upload logic.
There was a large chunk of code in sign/attest that was *largely* the same, and one instance was already leaning on the other for various utilities.  This change shares all of the common elements of this large chunk that were common with one small callback for handling the difference.

I'd have liked to have put this new helper into `cosign` alongside some of the other utilities, but it seems this would create a dependency cycle.

Signed-off-by: Matt Moore <mattomata@gmail.com>