Fix the postsubmit failure introduced by #795
The logs of a failure are here: https://github.com/sigstore/cosign/runs/3705988573#step:5:492

The problem here is that when I changed the return value to be the `name.Digest`, I was setting `lastHash` to the `v1.Layer` hash, not the `v1.Image` hash, so when the post-submit job tried to sign the resulting digest things exploded.

This change properly sets `lastHash` to the `v1.Image` hash, and uses a separate `layerHash` variable for the layer's digest for the download print statement we print.

Signed-off-by: Matt Moore <mattomata@gmail.com>