Allow `cosigned` to validate `Fulcio` signatures.
This change makes it so that when `cosigned` is enabled on a namespace, and no public key is provided the webhook will verify things against the Fulcio root.

The basic idea here is that this roughly matches the CLI where omitting the key will verify things against the Fulcio root instead.

Related: https://github.com/sigstore/cosign/issues/866

Signed-off-by: Matt Moore <mattomata@gmail.com>