Re-expose commands.
I was trying to pull in the recent upstream updates to a downstream repo that was invoking the CLI via `cli.Sign().Exec(...)` and this regressed in the cobra migration.  In a perfect world, we could pull in the CLI functionality without needing to go through `cobra.Command` to do it, but we're not there yet.  The alternative to `cli.Sign().Exec(...)` is `cli.SignCmd(...)`, which takes 11 arguments and growing each release.  The alternative fo `cli.Attest().Exec(...)` is `cli.AttestCmd(...)` with 9 arguments.  etc, etc, etc...

What's more, the above functions are also colocated with their cobra wrappers so we also can't save on not vendoring cobra yet, so it's a mystery to me why we'd force folks to not use this interface.

Fixes: https://github.com/sigstore/cosign/issues/880

Signed-off-by: Matt Moore <mattomata@gmail.com>