Pin crane dependency used in e2e tests
`go install` without a version specified installs the latest semver release of the tool. This means that a compromise of the go-containerregistry repo could result in arbitrary code running in this workflow, e.g., when e2e tests invoke `crane ls`.

With a version specified, even if the repo moves the `v0.6.0` tag to another commit, the Go module proxy will serve the previous v0.6.0 code.

v0.6.0 is the latest release of crane: https://github.com/google/go-containerregistry/releases

Signed-off-by: Jason Hall <jasonhall@redhat.com>