Add a test/example for signing using GitHub OIDC
This demonstrates building a container image consisting of a Go binary
(cmd/cosign in this case), pushing it to a registry, and signing it
using ambient GitHub OIDC credentials available in the GitHub Actions
build environment.

This should serve as a runnable/forkable demo, and also a test to guard
against breakages.

Because permission to push packages is not available to workflow runs
started by PRs, this action runs on pushes to main, and on a nightly
basis.

Signed-off-by: Jason Hall <jasonhall@redhat.com>