Hook up k8schain to verification.
This wires up k8schain to the verification path, in addition to the resolution path.

I was beating my head against the well wondering why my GKE identity configuration was wrong, but it was fine.  It wasn't using it because it was using the default keychain instead of k8schain!

Signed-off-by: Matt Moore <mattomata@gmail.com>