Allow disabling `verifySCT`.
This adds a new flag to the Fulcio options to allow disabling `verifySCT` on keyless signing paths.

The option is modeled after `InsecureSkipVerify` in `tls.Config` to convey the severity of disabling this verification, and the flag indicates that this is intended only for testing.

This is related to my work here: https://github.com/sigstore/fulcio/pull/216

Signed-off-by: Matt Moore <mattomata@gmail.com>