Return better errors from `cosigned`
This drops the `bool` on `valid()` so that when internal errors occur they can be surfaced back through the webhook to better indicate what problems the webhook is having.

To illustrate these better diagnostics, I've added unit tests that simulate errors looking up secrets and service accounts in k8schain.

Fixes: https://github.com/sigstore/cosign/issues/963

Signed-off-by: Matt Moore <mattomata@gmail.com>