Fix semantic bugs in attestation verifification.
DSSE requires that the payload types match, but the Attestation
spec requires that the payload type be the specific in-toto one. We
don't actually do anything with the content here other than dump it
in plain text, but semantically we're calling these attestations in the
CLI so we should enforce that part of the specification as well.

Signed-off-by: Dan Lorenc <lorenc.d@gmail.com>