Add TUF timestamp to OCI signature on sign
This adds the TUF timestamp to the Fulcio and Rekor
signers. Both are necessary since each relies on
TUF metadata. If both signers are used, the latter
one will overwrite the TUF timestamp.

I also added a basic mock Rekor client for tests.
A number of methods are not implemented yet.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>