Fix comparison in replace option for attestation
Currently the replace option compares the following when evaluating each
signature:
  r.predicateURI == payloadData["payloadType"]

This appears to be comparing the payload's predicateType (e.g a URI
"https://example.com/predicate" or "cosign.sigstore.dev/attestation/v1"
for custom to the signature's payloadType (which is always
"application/vnd.in-toto+json").

Fix this by making the replace option replace all attestations which
match the current predicate type.

Signed-off-by: Blake Burkhart <blake.burkhart@us.af.mil>