Improve json output of cosign verify
Separate the payload, cert and bundle output into separate fields,
rather than stuffing the latter two into the former.

I've also added the TUF timestamp to the output.

I've omitted the chain, because I'm not sure what fields in the
certificates are or aren't useful to include.

The annotations seem to be where all the former information (excluding
the payload) are actually stored, so I'm not including that.

Signed-off-by: Rob Best <robertbest89@gmail.com>