Add certificate chain flag for signing (#1656)
* Add certificate chain flag for signing

This allows users to pass their own certificate chain to include in the
OCI signature. The chain is checked for validity using the provided
certificate.

Also refactored the check for matching public keys using a method from
sigstore/sigstore, comparing the certificate's key with the provided
key. Also added this check when extracting the PKCS11 certificate.

Certificate chains must be PEM-encoded. I changed the text of the
certificate flag to also specify a preference for PEM encoding, but
didn't remove the code that handles DER encoding for backwards
compatibility.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Adding 3rd party licenses

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Added check for empty chain

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>