feat: add ability to override registry keychain
When importing cosign as a library directly in code, it is convenient
to have the option of providing the go-containerregistry keychain. This
allows consumers of cosign to call functions such as sign.SignCmd and
copy.CopyCmd and pass the registry creds in memory without having to
write the creds to a docker/oci auth config file first.

This commit adds a new type alias Keychain which maps to authn.Keychain.

Resolves: #1665

Signed-off-by: Michael Kucinski <MichaelKucinski@gmail.com>