cosign clean: Don't log failure if the registry responds with 404
remote.Delete expects either a 200 or 202, but if the tag doesn't exist
in the registry, some registries (seemingly ECR, maybe others) respond
with a 404, which just means there was nothing to delete.

This change also refactors logging to only log if the deletion was
successful, to avoid confusion if we can tell the tag didn't exist
before cleaning.

Signed-off-by: Jason Hall <jasonhall@redhat.com>