Fix handling of policy in verify-attestation (#1672)
* Fix sanity checks during policy processing

Previously, the code was comparing a predicate URI against a payload
type. This is never expected to match.

Modify the code to extract the actual predicate type from within the
payload so the verification can be done successfully.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* Fix rego policy verification

This commit addresses multiple issues when applying a rego policy
against the payload of an attestation.

1. If `data.signature.deny` evaluated to `true`, the policy verification
would pass. This is obviously unexpected. The code now looks for
`data.signature.allow` instead, and expects it to be `true`.

2. If a query result returned an undefined results, the policy
verification would pass. The code now explicitly checks for this
condition and ensure that if `ResultSet.IsAllowed()` returns `false`,
the policy verification also fails.

3. Improve error messages to assist user in defining correct variable
name and type.

4. Add unit tests to validate behavior and prevent breaking changes in
the future.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>