Make CLI flag for OIDC client secret take a path
Before, `--oidc-client-secret` secret was specified on the command line.
This runs the risk of leaking the secret to any other processes on the
same machine.

Now, the new flag `--oidc-client-secret-file` takes a path to a file
containing the OIDC secret.

This is a breaking change, but the flag is behind `COSIGN_EXPERIMENTAL=1`.

Signed-off-by: Zachary Newman <z@znewman.net>