Dont overwrite token set in keyOpts
This commit fixes a bug where an identity token set via options would
get overwritten if one of the providers returned one. Now, when an
identity token is available at the time of signing, the providers will
not be queried to generate a new one.

Thanks to Carlos Panato for finding the source of this bug that surfaced
in the Kubernetes image promoter.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>