Use bundle log ID to find verification key
Instead of iterating over all log verification keys, we use the log ID
in the bundle to find the correct key. This avoids iterating over all
trusted keys.

Note that the log ID is not signed. If the offline bundle is manipulated
so that the log ID does not resolve to a valid key, this will not result
in a denial of service attack, because cosign will fall back to fetching
the entry from the online log.

The equivalent change for CT entry verification has been made in another
PR.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>