Add privacy statement for PII storage
This statement has been added just before the OIDC code/normal flow,
since only in this flow we expect a user to be present. For the device
flow or when a token is provided, it's likely that Cosign is being run
in an automated environment.

Another option is to explicitly require that a user type 'y' or pass a
flag to bypass this check, but this will break existing usage.
Therefore, I simply added a sleep statement to pause to give users a
chance to read the message.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>