Check error on chain verification failure
This fails certificate verification at the correct step.

err was declared in the else scope, but was not checked before exiting
the conditional. Therefore, if a failure occurred, we didn't return
an error. However, later in the code, in verifyBlob, we do an extra
check for if the certificate is valid. Because we haven't set the root
and intermediate certs yet, cert verification fails.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>